working not good enghg
Eva i hope you don't see this because if you do it mean that filippo showed you this so here it goes.

hello baby girl, how you doing?
i was think of you , let me take you out for some dinner  and let me hit ( the dougie ) afterward and if not don't worry we can do that after the third date :) ;0